py_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain==0.3.2
langchain-community==0.3.1
langchain-core>=0.3.15,<0.4.0
langchain-text-splitters==0.3.2
langchain-openai==0.2.2
faiss-cpu==1.7.4
sentence-transformers==3.0.1
httpx==0.27.2
numpy==1.26.4
tenacity==8.5.0
pydantic==2.9.2

